requirements.txt
appnope==0.1.3
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
black==22.8.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
datatest==0.11.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.0
flake8==5.0.4
fonttools==4.37.4
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
stack-data==0.5.1
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
wcwidth==0.2.5
